Undo GWBK changes I made which serve no purpose.
