+++ src/main/kotlin/io/github/inductiveautomation/kindling/zip/views/gwbk/GatewayNetworkStatisticsRenderer.kt
@@ -2,7 +2,6 @@ package io.github.inductiveautomation.kindling.zip.views.gwbk
 
 import com.formdev.flatlaf.extras.FlatSVGIcon
 import com.formdev.flatlaf.extras.components.FlatTabbedPane
-import io.github.inductiveautomation.kindling.statistics.StatisticRenderer
 import io.github.inductiveautomation.kindling.statistics.categories.GatewayNetworkStatistics
 import io.github.inductiveautomation.kindling.statistics.categories.GatewayNetworkStatistics.IncomingConnection
 import io.github.inductiveautomation.kindling.statistics.categories.GatewayNetworkStatistics.OutgoingConnection
