+++ src/main/kotlin/io/github/inductiveautomation/kindling/zip/views/gwbk/OpcConnectionsStatisticsRenderer.kt
@@ -1,7 +1,6 @@
 package io.github.inductiveautomation.kindling.zip.views.gwbk
 
 import com.formdev.flatlaf.extras.FlatSVGIcon
-import io.github.inductiveautomation.kindling.statistics.StatisticRenderer
 import io.github.inductiveautomation.kindling.statistics.categories.OpcServerStatistics
 import io.github.inductiveautomation.kindling.utils.ColumnList
 import io.github.inductiveautomation.kindling.utils.FlatScrollPane
